requirements.txt
langchain
openai
python-dotenv
huggingface-hub
transformers
sentencepiece
cohere
replicate